+++ sphinxcontrib/osint/scripts/text.py
@@ -1,12 +1,13 @@
 # -*- encoding: utf-8 -*-
 """
-The text import scripts
+The text scripts
 ------------------------
 
 
 """
 from __future__ import annotations
 import os
+import argparse
 import sys
 from datetime import date
 import json
@@ -16,43 +17,20 @@ from sphinx.config import Config
 from sphinx.util.docutils import docutils_namespace
 
 from ..plugins.text import Text
+from . import parser_makefile, get_parser
 
 
 __author__ = 'bibi21000 aka Sébastien GALLET'
 __email__ = 'bibi21000@gmail.com'
 
-import argparse
 
 def get_parser_import(description='Description'):
     """Text import parser
     """
-    parser = argparse.ArgumentParser(
-        description=description,
-        )
-    parser.add_argument('--docdir', help="The documentation dir (where is the Makfile or make.bat)", default='.')
+    parser = get_parser(description=description)
     parser.add_argument('--delete_cache', help="Delete file in text_cache", action='store_true')
     parser.add_argument('textfile', nargs=1, help="The file to import in text store")
     return parser
-
-def parser_makefile(docdir):
-    sourcedir = None
-    builddir = None
-    if os.name == 'nt':
-        mkfile = os.path.join(docdir, 'make.bat')
-    else:
-        mkfile = os.path.join(docdir, 'Makefile')
-    if os.path.isfile(mkfile):
-        with open(mkfile, 'r') as f:
-            data = f.read()
-        lines = data.split('\n')
-        for line in lines:
-            if sourcedir is None and 'SOURCEDIR' in line:
-                tmp = line.split("=")
-                sourcedir = tmp[1].strip()
-            elif builddir is None and 'BUILDDIR' in line:
-                tmp = line.split("=")
-                builddir = tmp[1].strip()
-    return os.path.join(docdir, sourcedir), os.path.join(docdir, builddir)
 
 def main_import():
     parser = get_parser_import()
@@ -94,7 +72,7 @@ def main_import():
       "tags": None,
     }
 
-    Text.update(app, result)
+    Text.update(app, result, args.textfile[0])
 
     storef = os.path.join(sourcedir, app.config.osint_text_store, os.path.splitext(os.path.basename(args.textfile[0]))[0] + '.json')
     with open(storef, 'w') as f: